Add more modules, fix minor issues
* Implemented MangaHere
* Implemented MangaFox
* Implemented MangaReader
* moved modules to separate directory for separate deployment

+++ cbzdl/main.py
@@ -117,8 +117,8 @@ def parseArguments():
 
     parser = argparse.ArgumentParser(sys.argv, description="Download a comic")
     parser.add_argument("url", type=str, help="The URL of the comic to download")
-    parser.add_argument("-s", "--start", action="store", default=0, type=int, help="Minimum chapter to start from")
-    parser.add_argument("-e", "--end", action="store", default=9000, type=int, help="Maximum chapter to include (up to 9000)")
+    parser.add_argument("-s", "--start", action="store", default=0, type=float, help="Minimum chapter to start from")
+    parser.add_argument("-e", "--end", action="store", default=9000, type=float, help="Maximum chapter to include (up to 9000)")
     parser.add_argument("-d", "--delay", action='store', type=int, default=1, help="Delay to introduce during download (seconds)")
     parser.add_argument("-v", "--verbose", action='store_true', help="Verbose mode")
